requirements.txt
langchain>=0.2.0
langchain-google-genai
langgraph
faiss-cpu
python-dotenv